myblogproject/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
Django==3.2.4
django-cors-headers==4.3.1
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
openai==0.27.7
Pillow==9.5.0
PyJWT==2.8.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tqdm==4.65.0
urllib3==2.0.2
yarl==1.9.2
